rename seed/blogs.js to seed/posts.js and update to correct naming convention
