--- seed/blogs.js
@@ -1,35 +1,0 @@
-const db = require('../db/connection')
-const Blog = require('../models/blog')
-
-db.on('error', console.error.bind(console, 'MongoDB connection error:'))
-
-const main = async () => {
-  const blogs =
-    [
-      {
-        "title": "Product 001",
-        "imgURL": "https://slack-imgs.com/?c=1&o1=ro&url=https%3A%2F%2Fi.imgur.com%2FuqjoPURh.jpg",
-        "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-        "author": "130"
-      },
-      {
-        "title": "Product 002",
-        "imgURL": "https://imgur.com/Hm3hsLu",
-        "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-        "author": "100"
-      },
-      {
-        "title": "Product 003",
-        "imgURL": "https://imgur.com/Hm3hsLu",
-        "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-        "author": "140"
-      }
-    ]
-  await Blog.insertMany(blogs)
-  console.log("created blogs")
-}
-const run = async () => {
-  await main()
-  db.close()
-}
-run()
