+++ seed/posts.js
@@ -1,0 +1,35 @@
+const db = require('../db/connection')
+const Post = require('../models/post')
+
+db.on('error', console.error.bind(console, 'MongoDB connection error:'))
+
+const main = async () => {
+  const posts =
+    [
+      {
+        "name": "Product 001",
+        "imgURL": "https://slack-imgs.com/?c=1&o1=ro&url=https%3A%2F%2Fi.imgur.com%2FuqjoPURh.jpg",
+        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+        "price": "130"
+      },
+      {
+        "name": "Product 002",
+        "imgURL": "https://imgur.com/Hm3hsLu",
+        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+        "price": "100"
+      },
+      {
+        "name": "Product 003",
+        "imgURL": "https://imgur.com/Hm3hsLu",
+        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+        "price": "140"
+      }
+    ]
+  await Post.insertMany(posts)
+  console.log("created posts")
+}
+const run = async () => {
+  await main()
+  db.close()
+}
+run()
